[semver:major] Upgrade Slack to notify on failed deploys
With the current configuration, we only get Slack notifications about
successful deployments, despite having the notification steps in the
compiled CircleCI steps.

This is an attempt to enable notifications for all cases.

As part of this upgrade, Slack configuration needs to change.

- **old** `SLACK_WEBHOOK_URL`
- **new** `SLACK_ACCESS_TOKEN` through `moj-slack-notifications` context

Slack orb 4+ requires an application to be set up and a
`SLACK_ACCESS_TOKEN` variable. To enable this, we've installed an app
called `ci-deploy-to-cloud-platform` on the MoJ Slack instance.

The token is shared in the `moj-slack-notifications` CircleCI context.
Please see [the example notification workflow][ex] for a usage example.

[ex]: src/examples/slack_notification.yml