[semver:patch] Always pull base image on builds
We ran into some cases where the image build used yesterday's base image
and it tripped our vulnerability scanning.

Since we knew it was fixed already, it was a stale image. To avoid, we
can now pull the base image every time: if it's up-to-date, great,
otherwise it must contain something meaningful.

Overall, it makes the build more predictable.