Allow the built docker container to be used in subsequent jobs. (#49)
This allows the built container image to be saved to the workspace and
recalled in subsequent jobs, meaning that you can build your container,
use it in several other jobs (i.e. CVE scanning, tests etc) without
having to publish it to the container repository first.

As a result, I've also introduced a separate `publish_docker` job here
too.